+++ src/Commands.cpp
@@ -301,6 +301,7 @@ static SeqStrings gCommandNames =
     "CmdToggleHighlightFormFields\0"
     "CmdTogglePageBoxes\0"
     "CmdConvertPdfToImages\0"
+    "CmdToggleUniformPageWidth\0"
     "CmdNone\0"
     "\0";
 
@@ -590,6 +591,7 @@ static i32 gCommandIds[] = {
     CmdToggleHighlightFormFields,
     CmdTogglePageBoxes,
     CmdConvertPdfToImages,
+    CmdToggleUniformPageWidth,
     CmdNone,
 };
 
@@ -879,6 +881,7 @@ SeqStrings gCommandDescriptions =
     "Toggle Highlight Form Fields\0"
     "Toggle Page Boxes\0"
     "Convert PDF to Images...\0"
+    "Toggle Uniform Page Width\0"
     "Do nothing\0"
     "\0";
 // clang-format on