+++ tests/run-almost-all.ts
@@ -145,6 +145,7 @@ import { testit as issue5995 } from "./issue-5995.ts";
 import { testit as issue5997 } from "./issue-5997.ts";
 import { testit as issue4494 } from "./issue-4494.ts";
 import { testit as issue6001 } from "./issue-6001.ts";
+import { testit as issue5512 } from "./issue-5512.ts";
 
 export const tests: NamedTest[] = [
   // --- no Sumatra process -------------------------------------------------
@@ -284,6 +285,7 @@ export const tests: NamedTest[] = [
   ["issue-5997", issue5997],
   ["issue-4494", issue4494],
   ["issue-6001", issue6001],
+  ["issue-5512", issue5512],
 ];
 
 export async function testit(opts?: SuiteOptions): Promise<void> {